api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
httpx==0.25.1
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
reportlab==4.0.7

